requirements.txt
blinker==1.3
certifi==2018.1.18
chardet==3.0.4
cloud-init==17.2
command-not-found==0.3
configobj==5.0.6
cryptography==1.2.3
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11
django-braces==1.12.0
django-oauth-toolkit==1.0.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.7.7
gunicorn==19.7.1
idna==2.6
Jinja2==2.8
jsonpatch==1.10
jsonpointer==1.9
language-selector==0.1
MarkupSafe==0.23
oauthlib==2.0.7
Pillow==5.0.0
prettytable==0.7.2
psycopg2==2.7.3.2
py2neo==3.1.2
pyasn1==0.1.9
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.6.1
pyserial==3.0.1
python-apt==1.1.0b1+ubuntu0.16.4.1
python-debian==0.1.27
python-systemd==231
python3-openid==3.1.0
pytz==2018.3
PyYAML==3.11
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
ssh-import-id==5.5
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
whitenoise==3.2.1
